+++ src/ambition/version_details.cc
@@ -23,6 +23,12 @@ std::vector<VersionUpdates> versionDetails = {
           _("Automatically detect the language on Windows."),
         },
       },
+      {
+        .heading = _(HEADING_CHANGED),
+        .items = {
+          _("Base the display of the building guide on the Help setting.  Not shown when set to None; only the grid lines when set to Brief; and both highlighting and grid lines when set to Detailed."),
+        },
+      },
     },
   },
   {